[MOV] utm: split models in different files
Purpose
=======

This commit cleans the UTM module and uses a different file for each model
in order to clarify organization and ease reading. Each model belongs to
its file.

Task-2245823
