+++ addons/utm/models/utm_tag.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from random import randint
+
+from odoo import fields, models
+
+
+class UtmTag(models.Model):
+    """Model of categories of utm campaigns, i.e. marketing, newsletter, ..."""
+
+    _name = 'utm.tag'
+    _description = 'UTM Tag'
+    _order = 'name'
+
+    def _default_color(self):
+        return randint(1, 11)
+
+    name = fields.Char(required=True, translate=True)
+    color = fields.Integer(
+        string='Color Index', default=lambda self: self._default_color(),
+        help='Tag color. No color means no display in kanban to distinguish internal tags from public categorization tags.')
+
+    _sql_constraints = [
+        ('name_uniq', 'unique (name)', "Tag name already exists !"),
+    ]
